Remove the shared directory and move files to server.
