+++ .npmignore
@@ -1,7 +1,6 @@
 client/
 server/
 test/
-shared/
 docs/
 scripts/
 images/
