+++ server/src/lib/AuthenticationLevel.ts
