+++ server/src/lib/Method2FA.ts
@@ -1,0 +1,3 @@
+import Method2FA from "../../../client/src/types/Method2FA";
+
+export default Method2FA;
