+++ server/src/lib/authentication/Level.ts
@@ -1,3 +1,3 @@
-import { default as Level } from "../../../../shared/AuthenticationLevel";
+import { default as Level } from "../AuthenticationLevel";
 
 export { Level };
