+++ server/src/lib/routes/firstfactor/post.ts
@@ -2,13 +2,13 @@ import BluebirdPromise = require("bluebird");
 import express = require("express");
 import ErrorReplies = require("../../ErrorReplies");
 import { AuthenticationSessionHandler } from "../../AuthenticationSessionHandler";
-import UserMessages = require("../../../../../shared/UserMessages");
+import UserMessages = require("../../UserMessages");
 import { ServerVariables } from "../../ServerVariables";
 import { AuthenticationSession } from "../../../../types/AuthenticationSession";
 import { GroupsAndEmails } from "../../authentication/backends/GroupsAndEmails";
 import { Level } from "../../authentication/Level";
 import { Level as AuthorizationLevel } from "../../authorization/Level";
-import { BelongToDomain } from "../../../../../shared/BelongToDomain";
+import { BelongToDomain } from "../../BelongToDomain";
 import { URLDecomposer } from "../..//utils/URLDecomposer";
 import { Object } from "../../../lib/authorization/Object";
 import { Subject } from "../../../lib/authorization/Subject";
