+++ server/src/lib/storage/UserDataStore.ts
@@ -8,7 +8,7 @@ import { U2FRegistration } from "../../../types/U2FRegistration";
 import { TOTPSecret } from "../../../types/TOTPSecret";
 import { AuthenticationTraceDocument } from "./AuthenticationTraceDocument";
 import { IdentityValidationDocument } from "./IdentityValidationDocument";
-import Method2FA from "../../../../shared/Method2FA";
+import Method2FA from "../Method2FA";
 
 // Constants
