--- shared/Method2FA.ts
@@ -1,3 +1,0 @@
-import Method2FA from "../client/src/types/Method2FA";
-
-export default Method2FA;
